Adds optional strict parameter enforcement via @StrictParams.
If a resource is decorated with @StrictParams, it means that the number of parameters supplied to that method via either query params or form params must strictly match the number specified in the method.  This is an additional level of enforcement beyond @RequiredParam because it additionally does not allow the request to supply additional parameters that do not match to anything identified in the method.